Add audio

+++ app/js/prefabs/Player.js
@@ -82,14 +82,24 @@ YellowSidd.Player.prototype.update = function () {
     this.body.velocity.x = this.walking_speed;
     this.direction = "RIGHT";
     this.animations.play('walking'); // Play walking animation.
-    this.walking_sound.play(); // Play walking sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.walking_sound.play(); // Play walking sound.
+    }
+
     this.scale.setTo(1, 1);
   } else if (this.cursors.left.isDown && this.body.velocity.x <= 0) {
     // Move left
     this.body.velocity.x = -this.walking_speed;
     this.direction = "LEFT";
-    this.walking_sound.play(); // Play walking sound.
     this.animations.play('walking'); // Play walking animation.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.walking_sound.play(); // Play walking sound.
+    }
+
     this.scale.setTo(-1, 1);
   } else {
     // Stop.
@@ -101,7 +111,11 @@ YellowSidd.Player.prototype.update = function () {
   // Jump only if touching a tile.
   if (this.cursors.up.isDown && this.body.blocked.down) {
     this.body.velocity.y = -this.jumping_speed;
-    this.jump_sound.play(); // Play jump sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.jump_sound.play(); // Play jump sound.
+    }
   }
 
   // Dies if touches the end of the screen.
@@ -115,7 +129,11 @@ YellowSidd.Player.prototype.update = function () {
       // Start a timer which will call the shoot method in a loop.
       this.shoot();
       this.shoot_timer.start();
-      this.fireball_sound.play(); // Play jump sound.
+
+      // Play sound only if player left button sound as on mode.
+      if (PLAY_SOUND) {
+        this.fireball_sound.play(); // Play jump sound.
+      }
     }
   } else {
     this.shoot_timer.stop(false);
@@ -129,7 +147,11 @@ YellowSidd.Player.prototype.hit_enemy = function (player, enemy) {
     this.score += enemy.score; // Add points to score.
     enemy.kill(); // Kill enemy.
     player.y -= this.bouncing;
-    this.kill_enemy_sound.play(); // Play kill enemy sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.kill_enemy_sound.play(); // Play kill enemy sound.
+    }
   } else {
     this.die(); // Restart level.
   }
@@ -138,8 +160,12 @@ YellowSidd.Player.prototype.hit_enemy = function (player, enemy) {
 YellowSidd.Player.prototype.die = function () {
   "use strict";
   this.lives -= 1;
-  this.lost_heart_sound.play(); // Play lost heart sound.
   this.shooting = false;
+
+  // Play sound only if player left button sound as on mode.
+  if (PLAY_SOUND) {
+    this.lost_heart_sound.play(); // Play lost heart sound.
+  }
 
   if (this.lives > 0) {
     this.game_state.restart_level(); // Player lost 1 life, but still have more then 0, so restart level eventually checkpoint.
